requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0

# Deep learning
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0  # YOLOv11

# Data processing
pandas>=1.3.0
PyYAML>=5.4.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Coordinate transformations
pyproj>=3.2.0

# Progress bars
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Optional: GPU acceleration
# cupy>=9.0.0  # Uncomment if using CUDA

# Development
black>=21.0
flake8>=3.9.0
isort>=5.9.0
